fix: list bot credentials/configuration
